refactor: update feature imports to use shared components and hooks
